Remove unnecessary type
